Add a checkbox to remove the upload directory.

+++ server.nim
@@ -78,6 +78,7 @@ function test_json() {
   File 3: <input type="file" name="testfile-3" accept="text/*"><br /><br />
   Input 1: <input type="text" name="testfield-1" value="Test"><br /><br />
   Input 2: <input type="text" name="testfield-2" value="Test"><br /><br />
+  <input type="checkbox" name="remove_upload_dir" value="yes" checked> Remove Upload Directory<br /><br />
   <input type="submit">
 </form>
 
@@ -137,9 +138,12 @@ $1
           html.add("</ul>")
         html.add("</ul>")
         html.add("Upload Directory: $1" % uploadDir)
+        
+      if httpbody.formdata.hasKey("remove_upload_dir") and httpbody.formdata["remove_upload_dir"] == "yes":
+        removeDir(uploadDir)
+        html.add(" (Removed)")
 
       await req.respond(Http200, resform % html)
-      removeDir(uploadDir)
 
     elif req.url.path == "/form":
       let httpbody = await parseBody(req)
